+++ src/backend/SvgBackend.cc
@@ -3,6 +3,7 @@
 #include <cairo-svg.h>
 #include "../Canvas.h"
 #include "../closure.h"
+#include <cassert>
 
 using namespace v8;
 
@@ -13,7 +14,10 @@ SvgBackend::SvgBackend(int width, int height)
 
 SvgBackend::~SvgBackend() {
   cairo_surface_finish(surface);
-  if (_closure) delete _closure;
+  if (_closure) {
+    delete _closure;
+    _closure = nullptr;
+  }
   destroySurface();
 }
 
@@ -22,7 +26,8 @@ Backend *SvgBackend::construct(int width, int height){
 }
 
 cairo_surface_t* SvgBackend::createSurface() {
-  if (!_closure) _closure = new PdfSvgClosure(canvas);
+  assert(!_closure);
+  _closure = new PdfSvgClosure(canvas);
   surface = cairo_svg_surface_create_for_stream(PdfSvgClosure::writeVec, _closure, width, height);
   return surface;
 }
@@ -30,6 +35,7 @@ cairo_surface_t* SvgBackend::createSurface() {
 cairo_surface_t* SvgBackend::recreateSurface() {
   cairo_surface_finish(surface);
   delete _closure;
+  _closure = nullptr;
   cairo_surface_destroy(surface);
 
   return createSurface();